I used the map element in HTML to create different links for different element on the image posted on the web.
With the help of Photoshop I was able to get all the coordinates needed for this project.

I mad sure I used all the map shapes namely; rect, circle, poly, default.

After creating this project, I now fully understand how to use the map and where and when to use the different shapes